Hold join hero globe and routes until the globe is ready.
Inline SVG painted before the PNG, so dots and planes popped in first. Reveal both together from below once the image decodes.
